requirements_display.txt
adafruit-circuitpython-ssd1306
adafruit-blinka
digitalio
Pillow
requests
psutil
flask
